+++ src/Service/MailService.php
@@ -7,6 +7,7 @@ use ItkDev\MetricsBundle\Service\MetricsService;
 use Symfony\Bridge\Twig\Mime\TemplatedEmail;
 use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
 use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 
 final readonly class MailService
@@ -15,6 +16,7 @@ final readonly class MailService
         private MailerInterface $mailer,
         private MetricsService $metricsService,
         private string $fromAddress,
+        private string $fromName,
         private string $replyAddress,
         private string $defaultAddress,
     ) {
@@ -27,6 +29,8 @@ final readonly class MailService
      *   The recipient's email address
      * @param array $context
      *   The context for the email template
+     * @param string $refId
+     *   References header id (used to link mails)
      * @param string $subject
      *   The subject of the email. Defaults to 'Test mail from alert manager'.
      * @param string $htmlTemplate
@@ -36,10 +40,10 @@ final readonly class MailService
      *
      * @throws MailException
      */
-    public function sendEmail(string $to, array $context, string $subject = 'Test mail from alert manager', string $htmlTemplate = 'test.html.twig', string $textTemplate = 'test.txt.twig'): void
+    public function sendEmail(string $to, array $context, string $refId, string $subject = 'Test mail from alert manager', string $htmlTemplate = 'test.html.twig', string $textTemplate = 'test.txt.twig'): void
     {
         $email = (new TemplatedEmail())
-            ->from($this->fromAddress)
+            ->from(new Address($this->fromAddress, $this->fromName))
             ->to($to)
             ->cc($this->defaultAddress)
             ->replyTo($this->replyAddress)
@@ -49,6 +53,10 @@ final readonly class MailService
             ->htmlTemplate('mails/'.$htmlTemplate)
             ->context($context)
         ;
+
+        // Set references header to thread/link mails.
+        list($localPart, $domainPart) = explode('@', $this->fromAddress);
+        $email->getHeaders()->addTextHeader('References', $refId.'@'.$domainPart);
 
         try {
             $this->mailer->send($email);