+++ CHANGELOG.md
@@ -8,6 +8,16 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+- Added application name to subject.
+- Wrapped application name in device info with link to the application.
+- Added from name to mail address.
+- Added device EUI to device model.
+- Added 'References' header with EUI to mails.
+
+## [1.0.1] - 2024-11-13
+
+- Added `skipBasedOnAppEndDate` to checks.
+
 ## [1.0.0] - 2024-10-22
 
 - Bumped version